+++ actix-tls/CHANGES.md
@@ -2,9 +2,13 @@
 
 ## Unreleased - 2021-xx-xx
 * Remove `trust-dns-proto` and `trust-dns-resolver` [#248]
-* Use `tokio::net::lookup_host` for as simple and basic default resolver [#248]
-* Add `Resolve` trait for custom dns resolver. Add `Resolver::new_custom` function
-  to construct custom resolvers.[#248]
+* Use `tokio::net::lookup_host` as simple and basic default resolver [#248]
+* Add `Resolve` trait for custom dns resolver. [#248]
+* Add `Resolver::new_custom` function to construct custom resolvers. [#248]
+* Export `webpki_roots::TLS_SERVER_ROOTS` in `actix_tls::connect` mod and remove
+  the export from `actix_tls::accept` [#248]
+* Remove `ConnectTakeAddrsIter`. `Connect::take_addrs` would return 
+  `ConnectAddrsIter<'static>` as owned iterator. [#248]
 
 [#248]: https://github.com/actix/actix-net/pull/248